# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.CallFeedbackRequest:
    properties:
      comment:
        description: 可选评论
        example: 通话质量良好，声音清晰
        type: string
      issues:
        description: 问题描述
        example: 偶尔有一点延迟
        type: string
      rating:
        description: 1-5 星评分
        example: 4
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
...
  controllers.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      data: {}
      message:
        example: Invalid username or password
        type: string
    type: object
...
host: localhost:20033
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: A comprehensive door access management system with video calling capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ILock HTTP Service API
  version: "1.0"
paths:
...
  /calls:
    get:
      consumes:
      - application/json
      description: Get a list of all call records in the system, with pagination
      parameters:
      - description: Page number, default is 1
        in: query
        name: page
        type: integer
      - description: Items per page, default is 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Call Records
      tags:
      - CallRecord
  /calls/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific call record by ID
      parameters:
      - description: Call Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Call Record By ID
      tags:
      - CallRecord
  /calls/{id}/feedback:
    post:
      consumes:
      - application/json
      description: Submit quality feedback for a specific call record
      parameters:
      - description: Call Record ID
        in: path
        name: id
        required: true
        type: integer
      - description: Feedback information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CallFeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit Call Feedback
      tags:
      - CallRecord
  /calls/device/{deviceId}:
    get:
      consumes:
      - application/json
      description: Get call records for a specific device, with pagination
      parameters:
      - description: Device ID
        in: path
        name: deviceId
        required: true
        type: integer
      - description: Page number, default is 1
        in: query
        name: page
        type: integer
      - description: Items per page, default is 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Device Call Records
      tags:
      - CallRecord
  /calls/resident/{residentId}:
    get:
      consumes:
      - application/json
      description: Get call records for a specific resident, with pagination
      parameters:
      - description: Resident ID
        in: path
        name: residentId
        required: true
        type: integer
      - description: Page number, default is 1
        in: query
        name: page
        type: integer
      - description: Items per page, default is 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Resident Call Records
      tags:
      - CallRecord
  /calls/statistics:
    get:
      consumes:
      - application/json
      description: Get call statistics including total, answered, missed, etc.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Call Statistics
      tags:
      - CallRecord
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"